requirements.txt
Django==1.8.16
requests
pynacl
cryptography
requests[socks]
toml
